fix(03-01): numpy/pandas compat in to_H2OFrame and apply tests (VAL-04)
pyunit_to_H2OFrame.py jagged-array case: numpy 2.x requires explicit
dtype=object to construct ragged 2D arrays from lists of unequal length.

pyunit_apply.py h2o_to_float helper: newer pandas keeps int dtype
through np.ceil/np.floor while H2O ceil/floor returns float, so the
helper now coerces the pandas side too.

+++ h2o-py/tests/testdir_misc/pyunit_apply.py
@@ -16,7 +16,9 @@ def h2o_to_float(h2o, pd):
     to compare with Pandas results.
     :return:
     """
-    return (h2o.astype(float), pd)
+    # Newer pandas keeps int dtype through np.ceil/np.floor; H2O ceil/floor returns
+    # float, so coerce both sides to float to compare values without dtype skew.
+    return (h2o.astype(float), pd.astype(float))
 
 
 def pd_to_int(h2o, pd):
